requirements.txt
Pillow==6.1.0
dataclasses==0.6.0
dataclass_factory==2.2
six==1.12.0
